backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0
alembic>=1.14.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
